chore(cron): move cron package under update package
