+++ update/cron/jobs_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	flatbuffers "github.com/google/flatbuffers/go"
-	cronfb "github.com/qri-io/qri/cron/cron_fbs"
+	cronfb "github.com/qri-io/qri/update/cron/cron_fbs"
 )
 
 func CompareJobSlices(a, b []*Job) error {
